+++ src/meteole/forecast.py
@@ -9,7 +9,7 @@ import tempfile
 from abc import ABC, abstractmethod
 from functools import reduce
 from importlib.util import find_spec
-from typing import Any, Optional
+from typing import Any
 
 import pandas as pd
 import xarray as xr
@@ -149,7 +149,7 @@ class WeatherForecast(ABC):
         run: str | None = None,
         interval: str | None = None,
         coverage_id: str = "",
-        temp_dir: Optional[str] = None,
+        temp_dir: str | None = None,
     ) -> pd.DataFrame:
         """Return the coverage data (i.e., the weather forecast data).
 
@@ -166,7 +166,7 @@ class WeatherForecast(ABC):
                     raises an error if specified. Defaults to "P1D" for time-aggregated indicators such
                     as TOTAL_PRECIPITATION.
             coverage_id: An id of a coverage, use get_capabilities() to get them.
-            temp_dir (Optional[str]): Directory to store the temporary file. Defaults to None.
+            temp_dir (str | None): Directory to store the temporary file. Defaults to None.
 
         Returns:
             pd.DataFrame: The complete run for the specified execution.
@@ -411,7 +411,7 @@ class WeatherForecast(ABC):
     def _grib_bytes_to_df(
         self,
         grib_str: bytes,
-        temp_dir: Optional[str] = None,
+        temp_dir: str | None = None,
     ) -> pd.DataFrame:
         """(Protected)
         Converts GRIB data (in binary format) into a pandas DataFrame.
@@ -422,7 +422,7 @@ class WeatherForecast(ABC):
 
         Args:
             grib_str (bytes): Binary GRIB data as a byte string.
-            temp_dir (Optional[str]): Directory to store the temporary file. Defaults to None.
+            temp_dir (str | None): Directory to store the temporary file. Defaults to None.
 
         Returns:
             pd.DataFrame: A pandas DataFrame containing the extracted GRIB data,
@@ -474,7 +474,7 @@ class WeatherForecast(ABC):
         height: int | None,
         lat: tuple,
         long: tuple,
-        temp_dir: Optional[str] = None,
+        temp_dir: str | None = None,
     ) -> pd.DataFrame:
         """(Protected)
         Return the forecast's data for a given time and indicator.
@@ -486,7 +486,7 @@ class WeatherForecast(ABC):
             forecast_horizon (int): the forecast horizon in hours (how many hours ahead)
             lat (tuple): minimum and maximum latitude
             long (tuple): minimum and maximum longitude
-            temp_dir (Optional[str]): Directory to store the temporary file. Defaults to None.
+            temp_dir (str | None): Directory to store the temporary file. Defaults to None.
 
         Returns:
             pd.DataFrame: The forecast for the specified time.
@@ -625,7 +625,7 @@ class WeatherForecast(ABC):
         lat: tuple = FRANCE_METRO_LATITUDES,
         long: tuple = FRANCE_METRO_LONGITUDES,
         forecast_horizons: list[int] | None = None,
-        temp_dir: Optional[str] = None,
+        temp_dir: str | None = None,
     ) -> pd.DataFrame:
         """
         Get a combined DataFrame of coverage data for multiple indicators and different runs.
@@ -645,7 +645,7 @@ class WeatherForecast(ABC):
             lat (tuple): The latitude range as (min_latitude, max_latitude). Defaults to FRANCE_METRO_LATITUDES.
             long (tuple): The longitude range as (min_longitude, max_longitude). Defaults to FRANCE_METRO_LONGITUDES.
             forecast_horizons (list[int] | None): A list of forecast horizon values in hours. Defaults to None.
-            temp_dir (Optional[str]): Directory to store the temporary file. Defaults to None.
+            temp_dir (str | None): Directory to store the temporary file. Defaults to None.
 
         Returns:
             pd.DataFrame: A combined DataFrame containing coverage data for all specified runs and indicators.
@@ -681,7 +681,7 @@ class WeatherForecast(ABC):
         lat: tuple = FRANCE_METRO_LATITUDES,
         long: tuple = FRANCE_METRO_LONGITUDES,
         forecast_horizons: list[int] | None = None,
-        temp_dir: Optional[str] = None,
+        temp_dir: str | None = None,
     ) -> pd.DataFrame:
         """(Protected)
         Get a combined DataFrame of coverage data for a given run considering a list of indicators.
@@ -701,7 +701,7 @@ class WeatherForecast(ABC):
             lat (tuple): The latitude range as (min_latitude, max_latitude). Defaults to FRANCE_METRO_LATITUDES.
             long (tuple): The longitude range as (min_longitude, max_longitude). Defaults to FRANCE_METRO_LONGITUDES.
             forecast_horizons (list[int] | None): A list of forecast horizon values in hours. Defaults to None.
-            temp_dir (Optional[str]): Directory to store the temporary file. Defaults to None.
+            temp_dir (str | None): Directory to store the temporary file. Defaults to None.
 
         Returns:
             pd.DataFrame: A combined DataFrame containing coverage data for all specified runs and indicators.